+++ packages/base/index.js
@@ -21,14 +21,6 @@ module.exports = {
         // @see https://eslint.org/docs/rules/curly
         'curly': ['error', 'all'],
 
-        // - La clause `default` doit être placée en dernier dans les switchs.
-        // @see https://eslint.org/docs/rules/default-case-last
-        'default-case-last': ['error'],
-
-        // - Les paramètres avec valeur par défaut doivent être placés en dernier.
-        // @see https://eslint.org/docs/rules/default-param-last
-        'default-param-last': ['error'],
-
         // - Le nom des fonctions doit matcher le nom des variables / propriétés
         //   à laquelle elles sont assignées (e.g. `const myFunc = function myFunc() {};`).
         // @see https://eslint.org/docs/rules/func-name-matching
@@ -53,10 +45,6 @@ module.exports = {
         //
         // @see https://eslint.org/docs/rules/func-style
         'func-style': ['error', 'expression'],
-
-        // - Les retours à la ligne dans les arguments de fonction doivent être cohérents.
-        // @see https://eslint.org/docs/rules/function-call-argument-newline
-        'function-call-argument-newline': ['error', 'consistent'],
 
         // - Les fonctions dont le nom ne peut être déduit doivent être nommées explicitement.
         //   Ceci permet un debugging plus simple car les fonctions sont nommés dans la console développeur des navigateurs.
@@ -87,16 +75,6 @@ module.exports = {
             ],
             optionalDependencies: false,
         }],
-
-        // - Empêche l'utilisation de `import ... from` avec les exports CommonJS (`module.exports = ...`, etc.)
-        // https://github.com/benmosher/eslint-plugin-import/blob/1012eb951767279ce3b540a4ec4f29236104bb5b/docs/rules/no-import-module-exports.md
-        'import/no-import-module-exports': ['error', {
-            exceptions: [],
-        }],
-
-        // - Empêche l'import de paquets (présence d'un `package.json`) via des imports relatifs.
-        // https://github.com/benmosher/eslint-plugin-import/blob/1012eb951767279ce3b540a4ec4f29236104bb5b/docs/rules/no-relative-packages.md
-        'import/no-relative-packages': ['error'],
 
         // - L'indentation doit être de 4 espaces (pas de tabs).
         // @see https://eslint.org/docs/rules/indent
@@ -147,14 +125,6 @@ module.exports = {
             ignoreUrls: true,
         }],
 
-        // - Interdit les retours dans les constructeurs.
-        // @see https://eslint.org/docs/rules/no-constructor-return
-        'no-constructor-return': ['error'],
-
-        // - Vérifie que les `else if` ne contiennent pas de doublons.
-        // @see https://eslint.org/docs/rules/no-dupe-else-if
-        'no-dupe-else-if': ['error'],
-
         // - Empêche certains exports nommés qui pourraient porter à confusion:
         //   - `then`: Cet export causeraient beaucoup de confusion dans les imports dynamiques.
         // https://eslint.org/docs/rules/no-restricted-exports
@@ -173,40 +143,9 @@ module.exports = {
         // @see https://eslint.org/docs/rules/no-implicit-globals
         'no-implicit-globals': ['error'],
 
-        // - Interdit l'assignation sur des éléments importés.
-        // @see https://eslint.org/docs/rules/no-import-assign
-        'no-import-assign': ['error'],
-
-        // - Interdit les nombres littéraux qui font perdre de la précision.
-        //   (Voir la page ESLint pour plus de détails)
-        // @see https://eslint.org/docs/rules/no-loss-of-precision
-        'no-loss-of-precision': ['error'],
-
-        // - Interdit les séquences d'échappement \8 et \9 dans les chaînes.
-        // @see https://eslint.org/docs/rules/no-nonoctal-decimal-escape
-        'no-nonoctal-decimal-escape': ['error'],
-
-        // - Interdit le retour dans la création de promesses (retour qui est de toute façon ignoré).
-        // @see https://eslint.org/docs/rules/no-promise-executor-return
-        'no-promise-executor-return': ['error'],
-
-        // - Interdit les retours de donnée depuis les mutateurs (setters).
-        // @see https://eslint.org/docs/rules/no-setter-return
-        'no-setter-return': ['error'],
-
-        // - Interdit les boucles avec un corps qui n'autorisent qu'une itération.
-        // @see https://eslint.org/docs/rules/no-unreachable-loop
-        'no-unreachable-loop': ['error', { ignore: [] }],
-
-        // - Empêche l'utilisation du chaînage optionnel dans les contextes
-        //   ou `undefined` n'est pas permis.
-        // @see https://eslint.org/docs/rules/no-unsafe-optional-chaining
-        'no-unsafe-optional-chaining': ['error', {
-            disallowArithmeticOperators: true,
-        }],
-
-        // @see https://eslint.org/docs/rules/no-useless-backreference
-        'no-useless-backreference': ['error'],
+        // - Pas de propriétés privées non utilisées dans les classes.
+        // @see https://eslint.org/docs/rules/no-unused-private-class-members
+        'no-unused-private-class-members': ['error'],
 
         // - Les objets littéraux doivent avoir des retours à la ligne cohérents.
         // @see https://eslint.org/docs/rules/object-curly-newline
@@ -240,18 +179,10 @@ module.exports = {
             allowUnboundThis: true,
         }],
 
-        // @see https://eslint.org/docs/rules/prefer-exponentiation-operator
-        'prefer-exponentiation-operator': ['error'],
-
         // - Préfère l'utilisation des groupes de capture dans les regex plutôt
         //   que la récupérations des correspondances via leurs indexes.
         // @see https://eslint.org/docs/rules/prefer-named-capture-group
         'prefer-named-capture-group': ['error'],
-
-        // - Préfère les regex littérales (e.g. `/(.+)/u`) au `new RegExp('(.+)', 'u')`.
-        //   (sauf dans le cas ou le constructeur `RegExp` est utilisé avec des variables)
-        // @see https://eslint.org/docs/rules/prefer-regex-literals
-        'prefer-regex-literals': ['error'],
 
         // - Des "quotes" simples doivent être utilisées partout, sauf pour tout ce qui
         //   est chaîne de caractères "non-système" (templates, phrases lues, ...)