+++ packages/vue/index.js
@@ -82,6 +82,7 @@ module.exports = {
         // enforce that class methods use "this"
         // @see https://eslint.org/docs/rules/class-methods-use-this
         'class-methods-use-this': ['error', {
+            enforceForClassFields: true,
             exceptMethods: [],
         }],